Set up basic grid for projects section
I want to make this adjust down to one column if there isn't
enough space. It does that now, but it doesn't feel exactly right
as there is some empty space in the single column at times. Also
the cards can be covered by the right column of the main grid, so
would maybe need to set a min size on project columns. Also, there
is a gap to the left of the li's in the ul. Maybe because it is
an ul?